Moved useEntityPagination; js_only.
Small cleanup; js_only.
