+++ assets/src/application/ui/layout/entityList/filterBar/useEntityListFilterState.ts
@@ -1,6 +1,16 @@
 import { useReducer } from 'react';
 
-const useEntityListFilterState = () => {
+export interface EntityListFilterState {
+	searchText: any;
+	setGridView: any;
+	setListView: any;
+	setSearchText: any;
+	showEntityFilters: any;
+	toggleEntityFilters: any;
+	view?: 'grid' | 'list';
+}
+
+const useEntityListFilterState = (): EntityListFilterState => {
 	const initialState = {
 		perPage: 6,
 		searchText: '',
@@ -23,13 +33,6 @@ const useEntityListFilterState = () => {
 		});
 	};
 
-	const setPerPage = (perPage) => {
-		dispatch({
-			perPage,
-			type: 'SET_PER_PAGE',
-		});
-	};
-
 	const setSearchText = (searchText) => {
 		dispatch({
 			searchText,
@@ -49,11 +52,9 @@ const useEntityListFilterState = () => {
 	const view = state.view;
 
 	return {
-		perPage,
 		searchText,
 		setGridView,
 		setListView,
-		setPerPage,
 		setSearchText,
 		showEntityFilters,
 		toggleEntityFilters,
@@ -65,9 +66,6 @@ const reducer = (state, action) => {
 	const { perPage, searchText, view } = action;
 
 	switch (action.type) {
-		case 'SET_PER_PAGE':
-			return { ...state, perPage };
-
 		case 'SET_SEARCH_TEXT':
 			return { ...state, searchText };
